refactor: 턴 검사하는 기능을 담당하는 객체를 변경
