+++ src/test/java/chess/domain/ChessBoardTest.java
@@ -39,7 +39,7 @@ class ChessBoardTest {
 	void attack() {
 		List<Piece> pieces = new ArrayList<>(Arrays.asList(new King(Side.WHITE, new Position("d4")), new Pawn(Side.BLACK, new Position("c5"))));
 		ChessBoard newChessBoard = new ChessBoard(pieces);
-		newChessBoard.move(new Position("d4"), new Position("c5"), Side.WHITE);
+		newChessBoard.move(new Position("d4"), new Position("c5"));
 		assertThat(newChessBoard.getPieces())
 				.extracting("position")
 				.isEqualTo(Collections.singletonList(new Position("c5")));
@@ -48,23 +48,15 @@ class ChessBoardTest {
 	@Test
 	@DisplayName("지정한 포지션에 없는 말을 선택한 경우 에러 처리")
 	void moveWithNonePiece() {
-		assertThatThrownBy(() -> chessBoard.move(new Position("c3"), new Position("c4"), Side.WHITE))
+		assertThatThrownBy(() -> chessBoard.move(new Position("c3"), new Position("c4")))
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessage("입력에 해당하는 말이 없습니다.");
 	}
 
 	@Test
-	@DisplayName("자기 턴이 아닌 경우에 상대 말을 움직이면 에러 처리")
-	void moveWithOtherSidePiece() {
-		assertThatThrownBy(() -> chessBoard.move(new Position("a7"), new Position("a6"), Side.WHITE))
-				.isInstanceOf(IllegalArgumentException.class)
-				.hasMessage("본인의 말만 움직일 수 있습니다.");
-	}
-
-	@Test
 	@DisplayName("규칙에 맞게 움직이지 않으면 에러처리")
 	void moveWithOutRule() {
-		assertThatThrownBy(() -> chessBoard.move(new Position("a2"), new Position("a5"), Side.WHITE))
+		assertThatThrownBy(() -> chessBoard.move(new Position("a2"), new Position("a5")))
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessage("해당 위치로 기물을 옮길 수 없습니다.");
 	}
@@ -73,7 +65,7 @@ class ChessBoardTest {
 	@DisplayName("규칙에 맞게 공격하지 않으면 에러처리")
 	void attackWithOutRule() {
 		List<Piece> pieces = Arrays.asList(new King(Side.WHITE, new Position("d4")), new Pawn(Side.BLACK, new Position("a2")));
-		assertThatThrownBy(() -> new ChessBoard(pieces).move(new Position("d4"), new Position("a2"), Side.WHITE))
+		assertThatThrownBy(() -> new ChessBoard(pieces).move(new Position("d4"), new Position("a2")))
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessage("해당 위치로 기물을 옮길 수 없습니다.");
 	}
@@ -81,7 +73,7 @@ class ChessBoardTest {
 	@Test
 	@DisplayName("길이 막힌 경우 에러처리")
 	void moveWithBlock() {
-		assertThatThrownBy(() -> chessBoard.move(new Position("a1"), new Position("a3"), Side.WHITE))
+		assertThatThrownBy(() -> chessBoard.move(new Position("a1"), new Position("a3")))
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessage("해당 위치로 가는 길이 막혀있습니다.");
 	}
