+++ src/test/java/chess/domain/ChessStatusTest.java
@@ -10,7 +10,6 @@ import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class ChessStatusTest {
 	ChessStatus chessStatus;
